Initial release of basic python2 binding
This is the first release of a basic out of tree python2 binding. It's
not entirely in sync with the python3 (lacks arch_to_personality) but
otherwise should behave identically.

This binding is not officially supported by LXC upstream as part of
1.0.x nor do we have any plan to include it for any upcoming release. It
is just released as a convenience for those who unfortunately can't yet
use python3.

Thanks to Dwight Engen for the original patchset!

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>